Resolución del issue 2 + separación por módulos
